backend/requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.3.0
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
